requirements.txt
discord.py[voice]
python-dotenv
google-genai
voicevox-core
